Allow output to stdout to provide greater flexibility
This commit adds the capability for unzboot to write the decompressed
image directly to standard output (stdout).

If the specified output file is "-", the program will now use `fwrite`
to send the output to stdout. Otherwise, it will continue to use
`g_file_set_contents` to write to the specified file.

This provides greater flexibility for piping the output of unzboot
to other command-line utilities.

Signed-off-by: Enric Balletbo i Serra <eballetb@redhat.com>