Suppress unused parameter warnings in zalloc/zfree
The `zalloc` and `zfree` functions, used as callbacks for zlib,
have a first parameter (named `x`) that corresponds to the `opaque`
field in the `z_stream` structure. This parameter allows passing
custom context to the allocation functions.

In the current implementation, this `opaque` context is not used,
leading to compiler warnings about an unused parameter `x`.

This commit silences these warnings by explicitly casting the `x`
parameter to `(void)x;`. This indicates to the compiler that the
parameter is intentionally unused while maintaining the function
signature required by the zlib API.

Signed-off-by: Enric Balletbo i Serra <eballetb@redhat.com>