Improve ZSTD decompression error handling and reporting
The `unpack_efi_zboot_image` function previously checked for ZSTD
decompression errors by testing if the return value of `ZSTD_decompress`
was less than zero. While this might catch some errors, the canonical
way to check for ZSTD errors is to use the `ZSTD_isError()` function,
as `ZSTD_decompress` returns a `size_t` which can be an error code
or the decompressed size.

This commit updates the error checking for `ZSTD_decompress` to use
`ZSTD_isError()`. Additionally, it now uses `ZSTD_getErrorName()`
to print a more specific and informative error message to stderr
when a ZSTD decompression failure occurs. This improves the
diagnostics for users when encountering corrupted or problematic
ZSTD-compressed payloads within EFI zboot images.

Signed-off-by: Enric Balletbo i Serra <eballetb@redhat.com>