Direct gunzip error messages to stderr
The `gunzip` function was previously printing error messages to
standard output using `printf` and `puts`. This is generally
undesirable for command-line utilities, as error messages should
be directed to standard error (stderr) to allow for easier
redirection of normal program output.

This commit changes all error reporting within the `gunzip` function
to use `fprintf(stderr, ...)` ensuring that diagnostic messages
are sent to the appropriate stream. This improves the utility's
behavior and adherence to common command-line tool conventions.

Signed-off-by: Enric Balletbo i Serra <eballetb@redhat.com>